move pending components into its own file
